ETL custom reports

+++ src/Database/MysqlDbConnection.php
@@ -209,7 +209,7 @@ class MysqlDbConnection extends DbConnection
                     break;
                     
                 case FieldType::FLOAT:
-                    $fieldDef .= 'FLOAT';
+                    $fieldDef .= 'BIGINT';
                     break;
     
                 case FieldType::CHAR:
@@ -545,7 +545,7 @@ class MysqlDbConnection extends DbConnection
                     break;
                 case FieldType::FLOAT:
                     if (isset($rowData[$fieldDbName]) && is_numeric($rowData[$fieldDbName])) {
-                        $rowValues[] = (float) $rowData[$fieldDbName];
+                        $rowValues[] = $rowData[$fieldDbName];
                     } else {
                         $rowValues[] = 'null';
                     }
